backend/requirements.txt
fastapi
uvicorn
pydantic-ai
python-dotenv
httpx
pydantic
